updated controllers and tests
